Refactor Chatwoot context relay into global asset
