+++ resources/views/filament/hooks/chatwoot-context-relay.blade.php
@@ -1,0 +1,1 @@
+<livewire:chatwoot-context-relay />
