+++ resources/views/livewire/chatwoot-context-relay.blade.php
@@ -1,0 +1,1 @@
+<div class="hidden" aria-hidden="true"></div>
